requirements.txt
asgiref==3.7.2
backports.zoneinfo==0.2.1
Django==4.0.10
djangorestframework==3.13.1
pytz==2023.3
sqlparse==0.4.4
typing_extensions==4.7.1
whitenoise==6.0.0
djangorestframework-simplejwt==5.1.0
gunicorn==20.1.0
Pillow==9.0.1
flake8==4.0.1
drf-yasg==1.20.0
python-decouple==3.6
